Add TxFloodManager.{cpp,h} and move pull-mode logic there
